web/requirements.txt
boto3==1.28.76
Flask>=2
pymysql==1.1.0
redis==4.6.0
requests==2.31.0
SQLAlchemy==2.0.25
alembic==1.12.1
argon2-cffi==23.1.0
psycopg2-binary==2.9.9
